add videoPlayer testing
